s/Orma/OrmaCore/

+++ example/src/main/java/com/github/gfx/android/orma/example/Todo.java
@@ -1,15 +1,15 @@
 package com.github.gfx.android.orma.example;
 
+import com.github.gfx.android.orma.annotation.Column;
 import com.github.gfx.android.orma.annotation.Index;
 import com.github.gfx.android.orma.annotation.PrimaryKey;
-import com.github.gfx.android.orma.annotation.Column;
 import com.github.gfx.android.orma.annotation.Table;
 
 @Table
 public class Todo {
 
     @PrimaryKey
-    public int id;
+    public long id;
 
     @Column
     @Index
@@ -17,5 +17,4 @@ public class Todo {
 
     @Column
     public String content;
-
 }
